+++ Sources/Protocols/AppearanceProtocol.swift
@@ -2,7 +2,7 @@
 //  Copyright (c) 2017 Touch Instinct
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
-//  of this software and associated documentation files (the Software), to deal
+//  of this software and associated documentation files (the "Software"), to deal
 //  in the Software without restriction, including without limitation the rights
 //  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 //  copies of the Software, and to permit persons to whom the Software is
@@ -11,7 +11,7 @@
 //  The above copyright notice and this permission notice shall be included in
 //  all copies or substantial portions of the Software.
 //
-//  THE SOFTWARE IS PROVIDED AS IS, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 //  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 //  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 //  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
@@ -20,20 +20,9 @@
 //  THE SOFTWARE.
 //
 
-import UIKit
+ //Protocol which ensures that specific type can apply appearance to itself
+public protocol AppearanceProtocol {
+    associatedtype Appearance
 
-/// ViewModel for EmptyCell
-public final class EmptyCellViewModel: SeparatorCellViewModel {
-
-    let color: UIColor
-    let height: CGFloat
-
-    /// Returns configured ViewModel for cell
-    /// - parameter height: Height of cell
-    /// - parameter color: Fill color of cell
-    /// - returns: Configured ViewModel
-    public init(height: CGFloat, color: UIColor = .clear) {
-        self.color = color
-        self.height = height
-    }
+    func configure(appearance: Appearance)
 }